% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{exactTrajectories}
\alias{exactTrajectories}
\title{Query patients from the database according to the exact trajectories specified in GUI}
\usage{
exactTrajectories(connection, dbms, schema, svector, ivector)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{svector}{The names of the states which form the observed trajectory}

\item{ivector}{The moments in time when the state occurs}
}
\description{
Query patients from the database according to the exact trajectories specified in GUI
}
