% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{importTrajectoryData}
\alias{importTrajectoryData}
\title{#' Query patients from the database according to the loose trajectories specified in GUI
#'
#' @param connection Connection to the database (DatabaseConnector)
#' @param dbms The database management system
#' @param schema Schema in the database where the tables are located
#' @param svector The names of the states which form the observed trajectory
#' @export
looseTrajectories <- function(connection, dbms, schema, svector) {
if (length(svector) < 1) {
return(message("Vector length smaller than 1!"))
}
if (length(svector) == 1) {}
\usage{
importTrajectoryData(connection, dbms, schema, trajectories)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{trajectories}{The matches in the matching table}
}
\description{
\preformatted{  sql = loadRenderTranslateSql(
    dbms = dbms,
    sql = paste("SELECT subject_id FROM @schema.patient_trajectories WHERE STATE_LABEL = '",svector[1], "';", sep = ""),
    schema = schema
  )
  eligiblePatients <- DatabaseConnector::querySql(connection,
                                                  sql = sql)
}
}
\section{sql2 = loadRenderTranslateSql(}{

}

\section{dbms = dbms,}{

}

\section{sql = "SELECT * FROM @schema.patient_trajectories WHERE SUBJECT_ID IN (@eligiblePatients);",}{

}

\section{schema = schema,}{

}

\section{eligiblePatients = eligiblePatients$SUBJECT_ID}{

}

\section{)}{

}

\section{returnData <- DatabaseConnector::querySql(connection,}{

}

\section{sql = sql2)}{

}

\section{Create set with eligible patients}{
eligiblePatients <- unique(unlist(returnList))

######################
}

\section{Query the data}{

}

