% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajFunctions.R
\name{looseTrajectories}
\alias{looseTrajectories}
\title{Query patients from the database according to the loose trajectories specified in GUI}
\usage{
looseTrajectories(connection, dbms, schema, svector)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{svector}{The names of the states which form the observed trajectory}
}
\description{
Query patients from the database according to the loose trajectories specified in GUI
}
