% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajFunctions.R
\name{outputAll}
\alias{outputAll}
\title{Query all trajectories defined in the settings}
\usage{
outputAll(connection, dbms, schema, settings)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{svector}{The names of the states which form the observed trajectory}
}
\description{
Query all trajectories defined in the settings
}
