% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{removeBeforeDataset}
\alias{removeBeforeDataset}
\title{Remove all instances for patient which occur before selected state occurrence}
\usage{
removeBeforeDataset(dataset, selectedState)
}
\arguments{
\item{dataset}{Patient trajectory data}

\item{selectedState}{Selected state}
}
\description{
Remove all instances for patient which occur before selected state occurrence
}
