% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{importTrajectoryData}
\alias{importTrajectoryData}
\title{#' Query patients from the database according to the exact trajectories specified in GUI
#'
#' @param connection Connection to the database (DatabaseConnector)
#' @param dbms The database management system
#' @param schema Schema in the database where the tables are located
#' @param svector The names of the states which form the observed trajectory
#' @param ivector The moments in time when the state occurs
#' @internal
exactTrajectories <- function(connection, dbms, schema, svector, ivector) {
if (length(ivector) != length(svector)) {
return(message("Vector length not equal!"))
}
else if (!all(sort(ivector) == 1:length(ivector))){
return(c())
}
else {}
\usage{
importTrajectoryData(connection, dbms, schema, trajectories)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{trajectories}{The matches in the matching table}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sql = "select * from ("
# Check for two same states in a row
checker = 0
for (index in 1:length(ivector)) \{
  sql = paste(
    sql,
    "select distinct SUBJECT_ID
           from (SELECT SUBJECT_ID,
                        STATE_LABEL,
                        STATE_START_DATE,
                        ROW_NUMBER()
                        OVER (PARTITION BY SUBJECT_ID ORDER BY SUBJECT_ID, STATE_START_DATE) as s_index
                 FROM @schema.@table) i
           where s_index =",
    ivector[index],
    " and STATE_LABEL ='",
    svector[index],
    "'", sep = "")
}\if{html}{\out{</div>}}
}
\details{

}
