% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{removeBeforeDatasetDB}
\alias{removeBeforeDatasetDB}
\title{Remove all instances for patient which occur before selected state occurrence}
\usage{
removeBeforeDatasetDB(connection, dbms, schema, selectedState)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{selectedState}{Selected state}
}
\description{
Remove all instances for patient which occur before selected state occurrence
}
